Refactor screens to use CompanyShopStyles for improved dark mode support
